novo componente, passados para pasta web_prototipos
novo prototipo de componente de perfil,
dummy user data e news data passadas agora por props pelo "main"
