+++ code/web_prototipos/public/index.html
@@ -26,6 +26,7 @@
       You need to enable JavaScript to run this app.
     </noscript>
     <div id="root"></div>
+    <div id="profile"></div>
     <div id="news"></div>
     <!--
       This HTML file is a template.
