+++ code/web_prototipos/run.sh
@@ -1,0 +1,19 @@
+#!/usr/bin/env bash
+set -eo pipefail
+
+case $1 in
+  start)
+    # The '| cat' is to trick Node that this is an non-TTY terminal
+    # then react-scripts won't clear the console.
+    yarn start | cat
+    ;;
+  build)
+    yarn build
+    ;;
+  test)
+    yarn test $@
+    ;;
+  *)
+    exec "$@"
+    ;;
+esac
