+++ code/web_prototipos/src/Styles/index.css
@@ -1,0 +1,5 @@
+
+
+#profile{
+	float:right;
+}
